update the customer food screen and controller

+++ lib/features/owner/home/presentation/widgets/owner_menu_product.dart
@@ -8,7 +8,7 @@ class OwnerMenuProduct extends StatelessWidget {
   final String image;
   final String foodName;
   final String price;
-  final int offerPrice;
+  final double offerPrice;
   final String star;
   final VoidCallback? onCheckTap;
   final VoidCallback? onDeleteTap;
